Added List, Char builtins & compilation support for suspended computations.
The pipe examples from paper.fk now execute and produce the expected
results. Compilation of lists and strings have also been fixed to obtain the
correct syntax; '|' for two "element" lists, the head and the tail and ','
separator otherwise.

Additionally, a simplification step has been added to the abstract machine to
automatically concatenate lists of strings produced by the consumption phase
of execution. It should therefore not be necessary to have separate syntax for
string concatenation. The motivation behind this change is so that I can write
generic list operations that do the "right thing" in the case of lists of
characters.

I have yet to add a test case for character builtins.