Added support for unifying suspended comps with flexibles.
Required to change the unification procedure so both
flexible ability and flexible type variables are analysed
together. The change has actually simplified the unification code
but there may be redundancy in the cases of cmp (in unify), solve
and solveForEVar. Unfortunately removing the final line (at least
for cmp) does not yield a pattern matching coverage warning.

Also added code to unify the interface maps and their arguments.