Fixed compilation to Shonky to correctly discriminate data types.
I've also added a pretty-printer for the Val data type.

The pretty-printing logic is complicated by the need to distinguish
builtin lists and lists representing constructors of other data types.

It was further complicated by the associativity I had chosen when
compiling constructors and constructor patterns. I had used foldl
whereas it is more convenient to use foldr to peel off the top
element of the list structure.