Fix Issue #7: perform one unfolding of flex-evars in ALL cases.
Previously, unfolding only occurred in the case of unifying two
abilities which both contained a flexible ability/effect
variable. Unfortunately, issue #7 shows this to be unsound. So
unfolding is now performed in all cases.

Moreover, it is performed on either ability (rather than both or
none, as previously). Finally, this change also enforces that only
one unfolding is attempted before yielding to solveForEVar. The
previous version of unifyAb potentially did multiple unfoldings
before yielding. It is unclear whether it matters but no regression
tests have failed due to this change.