Added a debug output flag to enable prettier errors by default.
This change introduces a controversial hack using
unsafePerformIO. My Haskell-fu is not up to a better solution at
the moment, and I think even if this approach is unsafe the only
downside is that it debugging output will not be produced since by
default debugging output on types is disabled.

The motivation for this change was to allow the addition of failing
tests where the test outputs a string containg types which themselves
contain type variables (both rigid and flexible). The unique identifier
is ignored by default: making the test directives less brittle and the
output easier on a user's eyes.

Added an example failing test for effect leaking from data
constructors.