Fixed unification of an effect-parametric datatype with itself.
Added a test case to highlight the resolution of a missing case
for unifying abilities; equality where the ability variable is
flexible. Bug manifested as a result of unifying the collection
of effects of an effect-parametric datatype with itself leading
to an infinite loop continually generating flexible ability
variables.

Result: The coop-factored.fk example's zipQueue operator can now
be defined with a more general type.

Comment(s):
  * I find it peculiar that this is the only case (so far) that
highlighted unification was not testing for equality in the case
of flexibles. Are there other, simpler examples?
  * unifyItfMap seems to be performing a redundant check; in most
cases it is called with the intersections of the two maps. Doing
this for all calls would presumably mean there is no need to
unify the maps twice in the definition of unifyItfMap. However,
I hestiate to make this modification for the moment, particularly
because it is precisely this apparent redundancy that gave rise
to this unification bug.