Added Conor's smuggling effects example with up-to-date syntax.
The example was to demonstrate problems with previous versions of
the Frank type system which permitted a continuation to be captured
in one context and executed in another, smuggling an effect not
handled by the second context along the way. The current Frank system
does not allow this example by forcing the ambient and the ability
of an operator's peg to agree exactly. In other words, they
are unified.

Extensions to the Frank type system may one day relax this
unification constraint in favour of some form of subeffecting, but
the smuggle example is forbidden eternal.