Fixed bugs in test harness and smuggle example.
Test harness wasn't correctly checking for a expected failing test.

Smuggle example had a scoping issue with the parameter A in Either
being mistaken for the datatype A. This is still a TODO but for now
I've just renamed the datatype.

I've also added a comment to the smuggle example to highlight
where the error appears if we change the pegs to be open. The error
shows up in type of juggle whose second argument is either some type
or a suspended computation that can do what the ambient can do plus
stateful computations; in other words, Aborts are not allowed.