Fixed a bug identified by Sam in examples/two-readers.fk
The problem lay in type-checking a use of shift. For this type-check, we first need to compare the instances-to-be-shifted with the current ambient (because first need to be suffix of the latter). But for this, the ambient needs to be expanded first (resolve all flex. eff. vars.).

expandAb, which is used for expansion, is now maximally expanding eff. vars. Another place where the single-unfold-expandAb was used can be considered broken too (unifyAb), it is now also using the new expandAb.

Also added some more debug support.