Modified the syntax of shift and fixed a regression: coop-old.fk.
The new syntax of shift is much simpler and does not rely on
unification. A set of interface names is specified between chevrons. If
those interfaces occur within the ambient then they are temporarily
removed during the inference of the second argument to shift. It is
an error if these interfaces do not occur within the ambient or if the
interfaces have not been declared. The test case shift-error.fk exhibits
the former error condition.

The fix for coop-old.fk was to explictly instantiate the effect variable
in flip's argument so it could occur in an ambient with two Co effects.
This example highlights a situation where using shift is inappropriate
even if it leads to the correct typing.