Modified command inference to account for a flexible ambient.
The ambient may have been created as part of "articulation" or
construction of the correct type shape during typechecking a thunk,
for example. We extend inference for a command to check for this
case and generate a new ability containing the corresponding
interface which is unified with the current ambient.

I've added a test case to demonstrate we can now infer the types
of effectful thunks.