Fixed shadowing bug in Refinement pass.
The Refinement phase no longer checks argument lengths to operators,
constructors or commands. Neither does it determine whether a variable
is monomorphic or polymorphic. The phase does not have enough type
information to accurately perform these checks. Such checks are
deferred until the TypeCheck phase.

The main purpose of this commit was to resolve the shadowing issue
identified by let-scoping.fk unit test. Since the Refine phase also
checked argument lengths, it made sense to port this functionality
over to the new code in TypeCheck. However, this commit does not
include tests to check correct argument lengths are supplied to
commands and constructors (and if not, that an error is reported).

The changes highlighted a bug in one of the unit tests,
flex-ab-eq.fk, where it was supplying the incorrect number of
arguments to a continuation variable. The code for this unit test
has been preserved in a regression (incorrect-num-args-mono.fk)
and rectified in the original unit test.